Feedback changes: move new functions out of issue-functions and increese sleep time to 30 seconds
